Purple tubes, IR sensor, multipliers
The Link can push the yellow multipliers into the rock heap, so that frees up a pair. The first row cubes go to orange poms, and the second row cubes go to the area where we push the red and green poms. We can (almost) grab the purple tube, even got it a couple times. Also we are using the built-in create3 IR distance sensors, bump sensors, and the cliff(line following) sensors. We also use the arc turning to push red and green poms into the area. Paper bumper needed to be taken downfor the IR sensors, and it turns out that the create3 can push the orange poms without it.